requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.26.3
python-multipart==0.0.6